move tx delete validator
